telegram-bot-kfc/requirements.txt
python-telegram-bot==20.7
pyodbc==4.0.39
python-dotenv==1.0.0
pandas==2.0.3
openpyxl==3.1.2
python-dateutil==2.8.2
pytz==2023.3
